Restructure explorer components
