+++ src/components/Explorer/DropZone/icons.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+export const scan = (
+  <svg
+    className="fill-current"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    width="64"
+    height="64"
+  >
+    <path fill="none" d="M0 0h24v24H0z" />
+    <path d="M4.257 5.671L12 13.414 13.414 12 5.671 4.257A9.959 9.959 0 0 1 12 2c5.523 0 10 4.477 10 10s-4.477 10-10 10S2 17.523 2 12c0-2.401.846-4.605 2.257-6.329z" />
+  </svg>
+);
+
+export const addToPlaylist = (
+  <svg
+    className="fill-current"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    width="64"
+    height="64"
+  >
+    <path fill="none" d="M0 0h24v24H0z" />
+    <path d="M18 15l-.001 3H21v2h-3.001L18 23h-2l-.001-3H13v-2h2.999L16 15h2zm-7 3v2H3v-2h8zm10-7v2H3v-2h18zm0-7v2H3V4h18z" />
+  </svg>
+);
+
+export const newPlaylist = (
+  <svg
+    className="fill-current"
+    xmlns="http://www.w3.org/2000/svg"
+    viewBox="0 0 24 24"
+    width="64"
+    height="64"
+  >
+    <path fill="none" d="M0 0h24v24H0z" />
+    <path d="M11 11V5h2v6h6v2h-6v6h-2v-6H5v-2z" />
+  </svg>
+);
